pyproject.toml
[project]
name = "your-own-data-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "delta-spark>=3.3.0",
    "faker>=37.1.0",
    "ipykernel>=6.29.5",
    "pyspark>=3.5.5",
    "ruff>=0.11.2",
]
